Collapse regression test into old, flawed test. Try to improve performance of two slow tests.
